LayoutTests/platform/mac/accessibility/no-unknown-objects-when-title-attribute-present.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
   testRunner.dumpAsText();
</script>
</head>
<body>

<div tabindex=0 id="divgroup" summary="Hello">
<span>
 <span>
   Text
 </span>
</span>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test that a parent object has help text, it does not cause intermediate, unknown objects to also be returned through accessibility. The test is that the child of the div is the 'Text' object and not the span objects.");

    if (window.accessibilityController) {
        document.getElementById("divgroup").focus();
        var divgroup = accessibilityController.focusedElement;
        shouldBe("divgroup.childAtIndex(0).role", "'AXRole: AXStaticText'");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>

</body>
</html>
