safari-534.53-branch/LayoutTests/platform/mac/accessibility/no-unknown-objects-when-title-attribute-present-expected.txt

Text
This test that a parent object has help text, it does not cause intermediate, unknown objects to also be returned through accessibility. The test is that the child of the div is the 'Text' object and not the span objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS divgroup.childAtIndex(0).role is 'AXRole: AXStaticText'
PASS successfullyParsed is true

TEST COMPLETE
